Remove the pangeo_forge module.
We never built out full functionality for loading pangeo_forge data in xarray-beam, and unit tests for this module are now failing at HEAD:
https://github.com/google/xarray-beam/actions/runs/6017267663

We can bring this back in the future when we are actually using it.

Also update our ReadTheDocs config so that our docs build works again.

PiperOrigin-RevId: 561140404